requirements.txt
pytorch_lightning~=1.6.5
open3d~=0.16.0
opencv-python~=4.5.5.62
PyYAML~=5.4.1
joblib~=1.2.0
loguru~=0.6.0
natsort~=8.2.0
opencv-python~=4.5.5.62
scipy~=1.9.3
ismember~=1.0.2
hydra-core~=1.3
omegaconf~=2.3.0
rich~=13.3.1
torchmetrics~=0.11.1
pyrootutils~=1.0.4
pytest~=7.2.1
packaging~=22.0